Move Selenium scraper out of view code and into data.etls
